api/requirements.txt
httpx
fastapi
uvicorn
python-dotenv
azure-storage-blob
fastapi-azure-auth